refactor: extract CommandGroupItem
